[Patient Intake Sample Generalizes setExtension helper function
